workflow/email.html: false context's from (#3171/#2750)
context's `from` was changed to `_from` in commit

> 24ded0c53 2022-01-13 congkhacdung  *.py: {} to dict with key arguments (#2750)

to avoid a name conflict with the python keyword `from`.
However, the commit missed to change the
corresponding name in `workflow/emal.html`.

The commit at hand renames the context variable to
`from_` consistently across python code and TAL template.
It also drops some unneeded variables in the
function that sets the context dictionary up.

Whenever there is a name collission, we should
use an underscore *suffix*; a prefix is
only used to indicate internal variables,
which could be accessed from outside, but shouldn't.