workflow: drop `workflowstep_fileattachment` (CAUTION: Migration) (#3166)
Now that all instances of the workflowstep node class
have been removed from the database in the parent commit,
we can remove the node class itself from the codebase.

WARNING: When this commit is applied to a running mediatum instance,
no node of the given type must be left over in the database.
This can be accomplished by executing the corresponding
alembic script in the parent commit of the commit
at hand BEFORE APPLYING THE COMMIT AT HAND.
Otherwise mediatum will complain about
undefined node types in the database.