node_to_access_rule: trigger fctn must get inherited rules (#2349)
The trigger fctn on_node_to_access_rule_insert_delete is called if
rules are inserted or deleted for any not inherited rules.
But if the last non-inherited rule is delete the rule is first
removed from the node and then the inherited rules are added to
the node.
When the last rule is removed the trigger fctn is called but there
are no rules at the node which can be inherited to the children.
When the inherited rules are added to the node the trigger fctn
isn't called because there are only inherited rules at the node.

The solution is, that the trigger fctn must add the inherited rules
to the node if no non-inherited rules are left and then inherit
these to the children.