web.frontend.frame: simplify `style_hide_empty`-handling (#3234)
Just treat the attribute as boolean value
(i.e. treat at True if it is a non-empty string).

There is no need for a special value for the collections root:
It is likely never empty, and even then,
it could use its own `style_hide_empty` value.

Also don't ask for `dont_ask_children_for_hide_empty`,
as it is defined nowhere.