schma.mask.field: use TAL 'get_form_html' also for mask editor (#3223)
This commit changes the rendering code
for the admin's mask editor:
metafield-related mask items are now
also rendered with the TAL macro `get_form_html`
(as it is already done for the user's editor).
This means that each meta field's HTML code
is now enclosed in two <fieldset> elements.

In the mask editor, we set
the outer fieldset to disabled.
This avoids the browser getting confused by
required meta fields that would otherwise prevent
the admin from submitting changes of the mask.

static/js/metadata/htmlmemo.js
--
javascript handling for WYSIWYG editor