utils.mail.sendmail: permit encrypted smtp (#3341)
Sending mails from mediatum is now possible
via encrypted smtp connections.
For this, the admin has to configure the
connection in the config file accordingly.

In the config file, the [server] block got dropped.
It must be replaced by a [smtp-server] block with
these configuration keys:

* `host`: the ip address or host name of the smtp server
* `port`: the port to be used
* `encryption`: may be one of the fixed strings
  "tls" with default port = 465 and connection over an
   SSL encrypted socket or "starttls" with default port = 587
* `username`: used to authenticate with the server;
  if unset, the server may not require an authentication
* `password-file`: path to a file with the
  password used for authentification

`port` defaults to 465 if `encryption`
is set to "tls", 25 otherwise.