contenttypes/document: make thumb with pdftoppm instead of gm (#3360)
With the update to Ghostscript 10,
graphicsmagick alters blue colors awkwardly
when converting pdf pages into raster images.
This started, when nixpkgs updated to Ghostscript 10, see

> https://github.com/NixOS/nixpkgs/pull/225815

which we applied to mediatum in

> 87bca8ebc fc80374d4 2023-06-21 Nikolaus Kolbenschlag Update mediatum dependencies:
  nixpkgs-23.11pre496033.c1bca7fe84c (#3295)

It is unclear why this happened at all, since
Graphicsmagic claims to *not* use Ghostscript
(`gm version` says it's not using "gslib").

There seems to be no way this can be rectified by
changing compile-time configuration of graphicsmagic.

Using pdftoppm generated raster images with correct colors.