workflow/checkcontent: to textpage workflowstep (#3341)
We change checkcontent workflowsteps to textpage workflowsteps in preparation of
the workflowstep class being removed from the code in a follow-up commit.

WARNING: The alembic script in this commit must be applied before updating the code
to commits further ahead. Otherwise mediatum will complain about undefined node types
in the database.