workflow/checkcontent: drop this workflowstep (CAUTION: Migration) (#3341)
WARNING: When this commit is applied to a running mediatum instance,
no node of the given type must be left over in the database.
This can be accomplished by executing the corresponding
alembic script in the parent commit of the commit
at hand BEFORE APPLYING THE COMMIT AT HAND.
Otherwise mediatum will complain about
undefined node types in the database.