web-data/web-root/static/img: unify similar images (#3383)
The following new files replace files
that contain the same impression
(albeit not bit-wise identical,
possibly even with differing file formats).
For each of these entries,
the file on the left hand side replaces
all files on the right hand side.
If the remaining file is also renamed
in the course of the change,
the file's old name is denoted between two colons:

* padlock.png: archive.png: changepwd.gif
* attachment-paperclip.png: attach.png: editor/mediatum_attach.svg
* editor/edit-pen-paper.svg: editor/mediatum_edit_black.svg: editor/mediatum_edit.svg edit.gif edit.png edit_opt.png edit_spc.png
* editor/magnifying-glass.svg: editor/mediatum_search.svg: search.gif frontend.gif
* webtree/trashbin.svg: editor/mediatum_delete_black.svg: editor/mediatum_delete.svg
* webtree/check.gif: haken.gif
* workflow.gif: icons/workflow.gif
* transparent-pixel.gif: mark_empty.png
* pdf.gif: mmicon_pdf.gif
* printer.png: print_icon.png: print.png
* webtree/directory.png: directory.png: webtree/directory.gif
* webtree/plus.svg: nav/nav-plus.svg: webtree/plus.gif
* webtree/spinner.gif: wait_small.gif: webtree/spinner.gif wait.gif
* layout/tabs.png: layout/edit_body.jpg