web-data/web-root/static/img/: unify duplicate icons (#3383)
The following new files replace
files with bit-wise identical data.
For each of these entries,
the files in the right hand side list
are bit-wise identical and are replaced
by the single file on the left hand side:

* webtree/collection.gif: webtree/collection.gif webtree/bare_collection.gif contextmenu/collection.gif
* transparent-pixel.gif: emtyDot1Pix.gif s.gif
* layout/bg-banner.png: layout/bg_banner.png layout/edit_bg_banner.png
* layout/menuitem-background-gradient.gif: layout/itenmenucontainer.gif layout/edit_itenmenucontainer.gif
* webtree/trashbin.gif: webtree/trashicon.gif contextmenu/trashicon.gif
* webtree/directory.gif: webtree/directory.gif webtree/folder.gif webtree/bare_directory.gif
* sort-by-alphabet-az.png: az.png ordersel.png
* sort-by-alphabet-za.png: za.png

With dropping `ordersel.png`,
the file `az.png` is renamed to the more
meaningful `sort-by-alphabet-az.png`.
For consistency,
the file `za.png` is also renamed accordingly,
even though it does not replace another file.