ckeditor_config: add ckeditor.updateElement() on every change (#3391)
Ckeditor has normally its own eventhandler for the submit event where
chkeditor.updateElement() is called.
Because mediaTUM also uses a submit eventhandler it is not garenteed,
which eventhandler is called at first.
To avoid this problem the ckeditor eventhandler is disabled with:
  ckeditor.config.autoUpdateElement = false;
and ckeditor.updateElement() is called on every change.