workflow/workflow.py: simplify forward and forwardAndShow (#3429)
textpage workflow step is the only user of the optional
`forward` argument, but it also doesn't need it:
It just forwards the node by itself before
calling `forwardAndShow(forward=False)`.
We simplify that by just calling `forwardAndShow()`
not no longer forwarding separately.

With this, the optional `forward` argument
is no longer needed by any caller.
It is removed and also not forwarded further
down in the chain of function calls.