utils/uwsgi.py: add unique timer for all workers (#3429)
Registers a uwsgi red-black timer for the
signal with the given `signal_name`.
The timer needs a `timer_name`, which is used as follows:
If a timer of this name was already registered in uwsgi
(e.g. by another worker process), it is not registered again.
If a timer of this name was already registered
in the calling process, an error is raised.
It is also an error to register a timer with
a different signal name, or with a different interval.
This function may only be called with uwsgi loaded.